requirements.txt
aenum==2.2.3
attrs==19.3.0
beautifulsoup4==4.9.0
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
dialogflow==0.8.0
fbchat==1.9.6
Flask==1.1.2
google-api-core==1.17.0
google-auth==1.14.1
googleapis-common-protos==1.51.0
grpcio==1.28.1
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.18.3
paho-mqtt==1.5.0
pandas==1.0.3
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pymessenger==0.0.7.0
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
requests-toolbelt==0.9.1
rsa==4.0
six==1.14.0
soupsieve==2.0
urllib3==1.25.9
Werkzeug==1.0.1
setuptools>=40.3.0
